tmp_code_816dc52cd85b63e14fdc1bfbd422f060.py

# BEGIN: user added these matplotlib lines to ensure any plots do not pop-up in their UI
import matplotlib
matplotlib.use('Agg')  # Set the backend to non-interactive
import matplotlib.pyplot as plt
plt.ioff()
import os
os.environ['TERM'] = 'dumb'
# END: user added these matplotlib lines to ensure any plots do not pop-up in their UI
   from promptshield.sdk import PromptShieldSDK
   
   sdk = PromptShieldSDK()
   result = sdk.classify_prompt("What is the capital of France?")