requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
redis>=4.5.0
PyYAML>=6.0
transformers>=4.30.0
openai>=0.27.0
anthropic>=0.3.0
requests>=2.28.0
python-dotenv>=1.0.0
click>=8.1.0
streamlit>=1.22.0
